Function files tested
Files were tested and cleaned. Will go back to comment more
