+++ exploratory-data-analysis-master/functions/EDAreview.py
@@ -1,17 +1,23 @@
+import pandas as pd
+import numpy as np
+import operator
+
 #FUNCTION TO GET TOP 10 POSITIVE AND NEGATIVE NOUN CHUNKS
 #Step 1 - Get reviews file and group by business id and aggregate noun chunks column and polarity score column:
+df['polarity'] = sentiment['polarity']
+df['ngram'] = df['ngram'].apply(lambda x: x.tolist())
+df['polarity_list'] = df.apply(lambda row: [row['polarity']]*len(row['ngram']), axis=1)
 sentiment_chunks = df.groupby('business_id').agg({'ngram': 'sum', 'polarity_list': 'sum'})
-#Step 2 - Create empty columns to be populated with list of noun chunks and sentiment scores.
-#Process noun chunks to return list of top 10 positive and negative noun chunks for each business
-# (one column for sentiment and second column for list of noun chunks)
 
-# the get_sentiment_chunks function is throwing an error 
+#Step 2 - Create empty columns to be populated with list of noun chunks and sentiment scores.
+#Process noun chunks to return list of top 10 positive and negative noun chunks for each business (one column for sentiment and second column for list of noun chunks)
+
 sentiment_chunks['top_chunks'] = np.nan
 sentiment_chunks['top_sentiment'] = np.nan
 from collections import defaultdict
 def get_sentiment_chunks(df):
     for i in range(len(df)):
-        if (type(df['top_chunks'].iloc[i]) != list):
+        if (type(sentiment_chunks['top_chunks'].iloc[i]) != list):
             dicts = [ {df['ngram'].iloc[i][j] : df['polarity_list'].iloc[i][j] } for j in range(len(df['ngram'].iloc[i])) ]
             result = {}
             intermediate = defaultdict(list)
@@ -32,6 +38,24 @@ def get_sentiment_chunks(df):
             pass
     return df
 
+#FUNCTION TO GET AVERAGE STAR RATING OVER AVAILABLE TIME PERIOD FROM REVIEWS FILE
+import datetime as datetime
+def avg_star_rating(df):
+   df['date_time'] = pd.to_datetime(df['date_time'])
+   df['date_list'] = df.apply(lambda row: [row['date_time']], axis=1)
+   df['star_review_list'] = df.apply(lambda row: [row['star_review']], axis=1)
+   grouped = df.sort_values(by=['business_id', 'date_time'], ascending=True).groupby('business_id')\
+                               .agg({'date_list': 'sum', 'star_review_list': 'sum'})
+   grouped['star_review_list'] = grouped['star_review_list'].apply(lambda x: np.array(x))
+   grouped['date_list'] = grouped['date_list'].apply(lambda x: np.array(x, dtype='datetime64[s]'))
+   grouped['star_ave'] = grouped['star_review_list'].apply(lambda x: [x[np.digitize(list(range(0, len(x), 1)),\
+                         np.linspace(0, len(x), 11)) == i].mean() for i in range(1, \
+                         len(np.linspace(0, len(x), 11)))] if len(x) > 10 else x)
+   grouped['date_ave'] = grouped['date_list'].apply(lambda x: [x[np.digitize(list(range(0, len(x), 1)), \
+                         np.linspace(0, len(x), 11)) == i].view('i8').mean().astype('datetime64[s]') \
+                         for i in range(1, len(np.linspace(0, len(x), 11)))] if len(x) > 10 else x)
+   return grouped
+
 
 #FUNCTION FOR GETTING REVIEWS BY YEAR
 def reviews_by_year(df):
@@ -46,10 +70,3 @@ def get_review_dist(df):
    review_dist = review_dist.reset_index()
    return review_dist
 
-# RUN_ALL function for ease of use
-
-def run_all(df):
-  get_sentiment_chunks(df)
-  reviews_by_year(df)
-  get_review_dist(df)
-  return df
